.csv files used with headers removed
abellcluster2.csv and comacluster2.csv are the raw data from the SDSS database. filtered_a and filtered_c are the files produced by running the code in 2b on the abell and coma cluster data respectively. abelllums and comalums are the files saved from topcat after calculating the luminosity of each galaxy, used in question 2e.